requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
dataclasses==0.7
google-api-core==1.22.2
google-auth==1.22.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
idna-ssl==1.1.0
libcst==0.3.11
multidict==4.7.6
mypy-extensions==0.4.3
ngram==3.3.2
numpy==1.19.2
pandas==1.1.2
pdf2image==1.14.0
Pillow==7.2.0
proto-plus==1.10.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
xlwt==1.3.0
yarl==1.6.0
